New version, had to change whole project structure because of errors caused by bad package definitions.
