--- .github/workflows/deploy.yml
@@ -1,37 +1,0 @@
-name: Java Deployment
-
-on:
-  push:
-    branches: [master]
-
-
-jobs:
-  build-and-deploy:
-    runs-on: self-hosted
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Build with maven
-        run: mvn clean package -DskipTests
-
-      - name: Move JAR to deployments folder
-        run: |
-          rsync -avz --exclude=".git" /etc/actions-runner/runner-notification-app/_work/notification-service/notification-service/ /home/salgadoth/deployments/notification-application/
-
-      - name: Set environment variables
-        run: |
-          echo "MONGODB_USERNAME=${{ secrets.MONGODB_USERNAME }}" >> $GITHUB_ENV
-          echo "MONGODB_PASSWORD=${{ secrets.MONGODB_PASSWORD }}" >> $GITHUB_ENV
-          echo "GC_EMAIL_SENDER=${{ secrets. GC_EMAIL_SENDER }}" >> $GITHUB_ENV
-
-      - name: Deploy application
-        run: |
-          cd ~/deployments/notification-application/
-          pm2 stop notification-application || true
-          pm2 delete notification-application || true
-          pm2 start ecosystem.config.js
-
-      - name: Save PM2 process list and environment
-        run: |
-          pm2 save
