+++ .gitignore
@@ -4,3 +4,4 @@ tokens
 .idea
 application.yml
 **/META-INF/MANIFEST.MF
+**/**.iml
