+++ common-kafka/src/main/java/br/com/tsg/collections/CorrelationId.java
@@ -1,0 +1,20 @@
+package br.com.tsg.collections;
+
+import java.util.UUID;
+
+public class CorrelationId {
+    private final String id;
+
+    CorrelationId(String title) {
+        id = title + "(" + UUID.randomUUID().toString() + ")";
+    }
+
+    @Override
+    public String toString() {
+        return "CorrelationId{" + "id=" + id + '}';
+    }
+
+    public CorrelationId continueWith(String title) {
+        return new CorrelationId(id + "-" + title);
+    }
+}
