+++ common-kafka/src/main/java/br/com/tsg/deserializer/GsonDeserializer.java
@@ -1,0 +1,20 @@
+package br.com.tsg.deserializer;
+
+import br.com.tsg.adapters.MessageAdapter;
+import br.com.tsg.collections.KafkaMessage;
+import com.google.common.reflect.TypeToken;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import org.apache.kafka.common.serialization.Deserializer;
+
+import java.lang.reflect.Type;
+
+public class GsonDeserializer<T> implements Deserializer<KafkaMessage<T>> {
+    private final Gson gson = new GsonBuilder().registerTypeAdapter(KafkaMessage.class, new MessageAdapter<T>()).create();
+
+    @Override
+    public KafkaMessage<T> deserialize(String topic, byte[] data) {
+        Type type = new TypeToken<KafkaMessage<T>>(){}.getType();
+        return gson.fromJson(new String(data), type);
+    }
+}
