--- common-kafka/src/main/java/br/com/tsg/notification_service/ConsumerFunction.java
@@ -1,7 +1,0 @@
-package br.com.tsg.notification_service;
-
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-
-public interface ConsumerFunction<T> {
-    void consume(ConsumerRecord<String, KafkaMessage<T>> record) throws Exception;
-}
