--- common-kafka/src/main/java/br/com/tsg/notification_service/CorrelationId.java
@@ -1,22 +1,0 @@
-package br.com.tsg.notification_service;
-
-import java.util.UUID;
-
-public class CorrelationId {
-    private final String id;
-
-    CorrelationId(String title) {
-        id = title + "(" + UUID.randomUUID() + ")";
-    }
-
-    @Override
-    public String toString(){
-        return "CorrelationId{" +
-                "id='" + id + '\'' +
-                '}';
-    }
-
-    public CorrelationId continueWith(String title){
-        return new CorrelationId(id + "-" + title);
-    }
-}
