--- common-kafka/src/main/java/br/com/tsg/notification_service/GsonSerializer.java
@@ -1,14 +1,0 @@
-package br.com.tsg.notification_service;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import org.apache.kafka.common.serialization.Serializer;
-
-public class GsonSerializer<T> implements Serializer<T> {
-    private final Gson gson = new GsonBuilder().registerTypeAdapter(KafkaMessage.class, new MessageAdapter()).create();
-
-    @Override
-    public byte[] serialize(String topic, T data) {
-        return gson.toJson(data).getBytes();
-    }
-}
