--- common-kafka/src/main/java/br/com/tsg/notification_service/KafkaDispatcher.java
@@ -1,51 +1,0 @@
-package br.com.tsg.notification_service;
-
-import org.apache.kafka.clients.producer.*;
-import org.apache.kafka.common.serialization.StringSerializer;
-
-import java.io.Closeable;
-import java.io.IOException;
-import java.util.Properties;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
-
-public class KafkaDispatcher<T> implements Closeable {
-    private final KafkaProducer<String, KafkaMessage<T>> producer;
-
-    KafkaDispatcher() {
-        this.producer = new KafkaProducer<>(properties());
-    }
-
-    private Properties properties() {
-        var properties = new Properties();
-        properties.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, "127.0.0.1:29092");
-        properties.setProperty(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
-        properties.setProperty(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, GsonSerializer.class.getName());
-        properties.setProperty(ProducerConfig.ACKS_CONFIG, "all");
-        return properties;
-    }
-
-    void send(String topic, String key, CorrelationId id, T payload) throws ExecutionException, InterruptedException {
-        Future<RecordMetadata> future = sendAsync(topic, key, id, payload);
-        future.get();
-    }
-
-    private Future<RecordMetadata> sendAsync(String topic, String key, CorrelationId id, T payload) {
-        var value = new KafkaMessage<>(payload.getClass().getName(), id, payload);
-        var record = new ProducerRecord<>(topic, key, value);
-        Callback callback = (data, ex) -> {
-            if(ex != null) {
-                ex.printStackTrace();
-                return;
-            }
-            System.out.println("sucesso enviando " + data.topic() + ":::partition " + data.partition() + "/ offset " + data.offset() + "/ timestamp " + data.timestamp());
-        };
-
-        return producer.send(record, callback);
-    }
-
-    @Override
-    public void close() throws IOException {
-        producer.close();
-    }
-}
