--- common-kafka/src/main/java/br/com/tsg/notification_service/KafkaService.java
@@ -1,75 +1,0 @@
-package br.com.tsg.notification_service;
-
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.common.serialization.StringDeserializer;
-
-import java.io.Closeable;
-import java.io.IOException;
-import java.time.Duration;
-import java.util.Collections;
-import java.util.Map;
-import java.util.Properties;
-import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.regex.Pattern;
-
-public class KafkaService<T> implements Closeable {
-    private final KafkaConsumer<String, KafkaMessage<T>> consumer;
-    private final ConsumerFunction<T> parse;
-
-    public KafkaService(String groupId, String topic, ConsumerFunction<T> parse, Map<String, String> properties) {
-        this(parse, groupId, properties);
-        consumer.subscribe(Collections.singletonList(topic));
-    }
-
-    KafkaService(String groupId, Pattern topic, ConsumerFunction<T> parse, Map<String, String> properties) {
-        this(parse, groupId, properties);
-        consumer.subscribe(topic);
-    }
-
-    private KafkaService(ConsumerFunction<T> parse, String groupId, Map<String, String> props) {
-        this.parse = parse;
-        this.consumer = new KafkaConsumer<>(getProperties(groupId, props));
-    }
-
-    private Properties getProperties(String groupId, Map<String, String> overrideProperties) {
-        var properties = new Properties();
-        properties.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "127.0.0.1:29092");
-        properties.setProperty(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
-        properties.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, GsonDeserializer.class.getName());
-        properties.setProperty(ConsumerConfig.GROUP_ID_CONFIG, groupId);
-        properties.setProperty(ConsumerConfig.CLIENT_ID_CONFIG, UUID.randomUUID().toString());
-        properties.setProperty(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, "1");
-        properties.putAll(overrideProperties);
-
-        return properties;
-    }
-
-    public void run() throws IOException, ExecutionException, InterruptedException {
-        try(var deadLetter = new KafkaDispatcher<>()){
-            while(true){
-                var records = consumer.poll(Duration.ofMillis(100));
-                if(!records.isEmpty()){
-                    System.out.println("Encontrei " + records.count() + " registros");
-                    for(var record: records) {
-                        try {
-                            parse.consume(record);
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                            var message = record.value();
-                            deadLetter.send("NEW_MESSAGE_DEAD_LETTER", message.getId().toString(),
-                                    message.getId().continueWith("DeadLetter"),
-                                    new GsonSerializer().serialize("", message));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @Override
-    public void close() {
-        consumer.close();
-    }
-}
