+++ email-service/src/main/java/br/com/tsg/collections/CorrelationId.java
@@ -1,0 +1,9 @@
+package br.com.tsg.collections;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class CorrelationId {
+    private String id;
+}
