+++ email-service/src/main/java/br/com/tsg/collections/RabbitMessage.java
@@ -1,0 +1,10 @@
+package br.com.tsg.collections;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class RabbitMessage {
+    private CorrelationId correlationId;
+    private Message payload;
+}
