--- email-service/src/main/java/br/com/tsg/notification_application/collections/CorrelationId.java
@@ -1,10 +1,0 @@
-package br.com.tsg.notification_application.collections;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-class CorrelationId {
-  private String id;
-}
