--- email-service/src/main/java/br/com/tsg/notification_application/collections/RabbitMessage.java
@@ -1,15 +1,0 @@
-package br.com.tsg.notification_application.collections;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RabbitMessage {
-    private CorrelationId correlationId;
-    private Message payload;
-}
-
-
-
