--- email-service/src/main/java/br/com/tsg/notification_application/exceptions/EmailSendingException.java
@@ -1,12 +1,0 @@
-package br.com.tsg.notification_application.exceptions;
-
-public class EmailSendingException extends Exception{
-    public EmailSendingException(String message){
-        super(message);
-    }
-
-    public EmailSendingException(String message, Throwable cause){
-        super(message, cause);
-    }
-}
-
